Fixed the C4267 warning
Turns out it is much better to use size_t whenver you index something, which I do quite often.

+++ source/engine/DepthBuffer.cpp
@@ -5,13 +5,13 @@
 extern RenderContext renderContext;
 extern DeviceContext deviceContext;
 
-DepthBuffer::DepthBuffer(UINT width, UINT height, UINT textureIndex, UINT dsvDescriptorIndex, const char* name)
+DepthBuffer::DepthBuffer(UINT width, UINT height, size_t textureIndex, size_t dsvDescriptorIndex, const char* name)
 	: width(width), height(height), textureIndex(textureIndex), dsvDescriptorIndex(dsvDescriptorIndex)
 {
 	strcpy_s(this->name, name);
 }
 
-UINT DepthBuffer::GetDescriptorIndex()
+size_t DepthBuffer::GetDescriptorIndex()
 {
 	return dsvDescriptorIndex;
 }
